refector(mentor): Dropdown, Nav-ProfileCard 영역 외 클릭 시 닫기 이벤트 유틸리티 함수로 변경
